Refactor priima software
in order to remove a lot of stuff that was needed for the research
project, but is no longer needed now. A bunch of new parameters make the
program nicer to use from the command line and the docker image is
updated such that it can be build from scratch without any preexisting
files in the program directory.