Create landmask shapefiles during image build
because these shapefiles are needed for the program to be able to
landmask the ground control points. For this purpose, OSM land polygons
that are already split into small chunks are downloaded and processed
at build time. A multi-stage build is introduced such that the larger
geopandas library does not need to be part of the final docker image.
The EPSG of the shapefiles is 4326 (lon/lat) as were the shapefiles from
the shapefiles used previously. This should be transformed to projected
coordinates eventually.