Update some geo packages
as shapely and other geo packages were outdated, their version numbers
are increased. The bump in shapely version is especially important
because we're planning on using the STRTree soon.
The complicated install of cartopy can be removed as we've dismissed
local installation in favor of a docker-only approach, thus we always
have the fitting GDAL and geos versions to install the higher cartopy
version numbers.