Raise error on invalid TPS output
the TPS implementation used here seems to be unstable, i.e. often times
an array full of zeros is returned instead of a warped image array. Up
until now these runs went unnoticed (until looking at the imagery).
Hence an error is raised to inform users. If this error is raised it can
be a good idea to either decrease the image size (via
--output-resolution) or the number of GCPs (via --num-gcps).